requirements.txt
fastapi==0.115.0
authlib==1.3.2
httpx>=0.28.1
python-dotenv==1.0.1
jinja2==3.1.4
itsdangerous==2.2.0
starlette==0.38.6
supabase==2.10.0
python-multipart==0.0.12
